feat(theory): Added experimentation script "overtones on the guitar".
The script allows to compute the site of any fret on the guitar neck
(even femto-tonal ones!). And, reversely, the fret number associated
with any site. By that it is possible to compute the frets associated
with each node in a given overtone index. The script also includes
a function for that.
.
The script is dedicated to Russian musician Marina Krupkina, kindling
the train of thought leading up to it by her video on overtones,
even though the original motiviation of that video is a different one.